Gave navigation bar a binded welcome message
Made the top of the navigation bar say welcome and the current users username. Also readjusted the binding of the navbar so its data context is the mainviewmodel instead of the navigator on the mainviewmodel. This way I could access the Authenticator on the mainviewmodel to bind to the current user.