Began creating UI for Projects Page
Began creating the UI for the projects view. I realized that it makes a lot more sense for me to create the views in order that they would actually be used (Project then ticket) so that I can actually create a project and then create a ticket for an existing project.