Created Projects View and made it so you can create new projects.
I started the UI for the projects page and the create new projects page. It works fine, but for some reason the database won't save any changes I make to the ProjectUsers ICollection on the User entity. When I create a new project it correctly adds the project and projectuser to the database, but when I exit out of the program, the changes to the ProjectUsers ICollection are gone.