Fixed Issue where ICollection wasn't being saved to database
I fixed the issue where the ProjectUsers ICollection didn't appear to be saving to the database. The issue was that I was not using eager loading (.Include() method when getting user entity) so I was not properly retrieving the collection. The Projects now are correcly displayed in the listbox on the projects view.