Worked on database updates when editing Ticket Details
I styled the Ticket title textbox to make it so it's highlighted when hovered over and then editable when clicked on. I also made it so that the database is automatically updated when the user changes the ticket tile in the view. However, I get a thread concurrency issue if I type too fast, so I need to find a way to fix that.