Fixed Ticket Status display having no spaces in TicketsView
I fixed the Ticket Status display having no spaces by added a helper property in the ticket class that used the StatusOptionsRetriever class to convert the status enum into a string and return that instead. I also moved the StatusOptionsRetriever class in BugTracker.Domain so that the ticket class wouldn't have to reference BugTracker.Wpf.