Got Search Bar to Work and Added EmailToIconConverter
I got the AddUserToProject search bar to work correctly (still need to put it in a popup so it doesn't change the size of the control). I also created a EmailToIconConverter so that when you search for a user in the search box, there is an icon next to the text that indicates whether it's an email or a username.