Changed the styling of the Bulletin Boards in ProjectDetailsPageView
Changed the styling of the ticket bulletin boards in the ProjectDetailsPageView. I used a uniform grid instead of how I was doing it before and this fixed the issue of the listbox not becoming scrollable. However, there is still an issue where I cant hover and scroll, I need to click the scrollbar and do it which is annoying. Still need to fix that.