Mostly fixed the Ticket Bulletin Scrolling Bug
Made it so that the user can scroll through the list of tickets in the ticket bulletins. However, the spacing makes it so that if one list is long, the create task buttons on all of the bulletins stay at the bottom. This is fine for now, but I'd like to fix this later.