Successfully created a Host for application
Created a host for application that configures app settings, etc.. Also switched the lifetime of multiple services to scoped because I learned that scoped meant a new instance per http request, not a new instance every time its injected. This means that things that are specific to each user such as the Authenticator because it holds the CurrentUser, and Navigator because it holds the CurrentViewModel.