Reformatted most of the WPF application to use Api Service
Reformatted most of the WPF application to use the api service instead of directly accessing the DbContext itself. I ran into some issues because the Data Transfer Objects no longer contain the navigation properties, only the ids of the navigation properties, so I will have to reformat the code to fit this new style.