Fixed main issues in TicketDetailsPage
Fixed all of the big issues in the Ticket Details Page after updating to using the api. The functionality of the page works exactly as intended, but there is a bug where if the User first sets an Assignee on a ticket, the assignee combo box does not display the name of the assignee, but after exiting the page and coming back in it does display the name. I will fix that issue after fixing the rest of the big issues.