Fixed issues with assignee and reporter combo boxes on TicketDetailsPage
Fixed issues with the assignee and reporter combo boxes. I was originally automatically updating the database when the user changed the value of the assignee or reporter combo box, but this led to issues with the update ticket command. I instead added an option for the user to confirm or cancel changes which got rid of the issues that stemmed from updating automatically after selecting a new value. The ticket details page now works exactly as intended.