Fixed issues with adding users to projects
Fixed issues with adding users to projects. Also made it so that pages that access the database disable user input while a database call is being made. There seems to suddenly be an issue with creating tickets where the Create ticket api endpoint doesn't return the correct values, so that needs to be fixed next.