Created custom authorization for api
Created custom authorization middleware and attribute for private api endpoints. The custom Authorization Middleware checks if the endpoint has the [CheckAuthorization] attribute and, if so, checks to see if there is a valid session with a session id that matches the session id in the cookie that was sent with the request.