Finished setting up Session Authentication
Switched authentication method from JWT to Session Authentication with secure cookies. Created middleware that checks to see if the client has a sessionId cookie and, if so, authenticates the session in the database to make sure the client is who they say they are and their session hasn't expired. Upon logging in, the Response returns an httponly cookie with the sessionId and that cookie is then automatically sent with every request. Logging out deletes the session from the database and the cookie.