Refactored Api for Tickets to be included in the ProjectDto
Fixed create ticket endpoint so it now correctly creates tickets with only the necessary information. TicketColumn comonent on the client now correctly filters and displays tickets in the correct columns on the project board. Also created an EnumToDisplayName converter so the Status and Priority enums are serialized as their display name.