Created TicketCard Component
Created a ticket card component to display ticket details such as title, assignee, and priority. Also acts as a button to open up the EditTicketModal. Also refactored the ticket.ts class on the client to make the author and assignee properties of type ProjectParticipant instead of having all the user properties directly on the ticket class.