Set up Refresh Tokens
Set up refresh token functionality in the application. Upon logging in, a refresh token is provided an stored in an httponly cookie. I created a Refresh endpoint in the api that provides the user with a new auth token, given that their current auth token is still valid. On the client, I made it so that there is a timer in the userStore that automatically refreshes the users auth token 60 seconds before it expires.